Move value objects from Fetch namespace #67
They should be reusable and not tied down to Fetch-ing
