+++ tests/Fetch/Branch/GitHubBranchTest.php
@@ -9,16 +9,16 @@ use Devboard\GitHub\Commit\Author\GitHubCommitAuthorEmail;
 use Devboard\GitHub\Commit\Author\GitHubCommitAuthorName;
 use Devboard\GitHub\Commit\Committer\GitHubCommitCommitterEmail;
 use Devboard\GitHub\Commit\Committer\GitHubCommitCommitterName;
+use Devboard\GitHub\Commit\GitHubCommitAuthor;
+use Devboard\GitHub\Commit\GitHubCommitAuthorDetails;
+use Devboard\GitHub\Commit\GitHubCommitCommitter;
+use Devboard\GitHub\Commit\GitHubCommitCommitterDetails;
 use Devboard\GitHub\Commit\GitHubCommitDate;
 use Devboard\GitHub\Commit\GitHubCommitMessage;
 use Devboard\GitHub\Commit\GitHubCommitSha;
-use Devboard\GitHub\Fetch\Branch\GitHubBranch;
-use Devboard\GitHub\Fetch\Commit\GitHubCommit;
-use Devboard\GitHub\Fetch\Commit\GitHubCommit as Commit;
-use Devboard\GitHub\Fetch\Commit\GitHubCommitAuthor;
-use Devboard\GitHub\Fetch\Commit\GitHubCommitAuthorDetails;
-use Devboard\GitHub\Fetch\Commit\GitHubCommitCommitter;
-use Devboard\GitHub\Fetch\Commit\GitHubCommitCommitterDetails;
+use Devboard\GitHub\GitHubBranch;
+use Devboard\GitHub\GitHubCommit;
+use Devboard\GitHub\GitHubCommit as Commit;
 use Devboard\GitHub\Repo\GitHubRepoFullName as RepoFullName;
 use Devboard\GitHub\Repo\GitHubRepoName;
 use Devboard\GitHub\User\GitHubUserApiUrl;
@@ -30,7 +30,7 @@ use Devboard\GitHub\User\GitHubUserLogin;
 use Devboard\GitHub\User\Type\User;
 
 /**
- * @covers \Devboard\GitHub\Fetch\Branch\GitHubBranch
+ * @covers \Devboard\GitHub\GitHubBranch
  * @group  unit
  *
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
