+++ tests/Fetch/Commit/GitHubCommitCommitterTest.php
@@ -6,9 +6,9 @@ namespace tests\Devboard\GitHub\Fetch\Commit;
 
 use Devboard\GitHub\Commit\Committer\GitHubCommitCommitterEmail;
 use Devboard\GitHub\Commit\Committer\GitHubCommitCommitterName;
+use Devboard\GitHub\Commit\GitHubCommitCommitter;
+use Devboard\GitHub\Commit\GitHubCommitCommitterDetails;
 use Devboard\GitHub\Commit\GitHubCommitDate;
-use Devboard\GitHub\Fetch\Commit\GitHubCommitCommitter;
-use Devboard\GitHub\Fetch\Commit\GitHubCommitCommitterDetails;
 use Devboard\GitHub\User\GitHubUserApiUrl;
 use Devboard\GitHub\User\GitHubUserAvatarUrl;
 use Devboard\GitHub\User\GitHubUserGravatarId;
@@ -18,7 +18,7 @@ use Devboard\GitHub\User\GitHubUserLogin;
 use Devboard\GitHub\User\Type\User;
 
 /**
- * @covers \Devboard\GitHub\Fetch\Commit\GitHubCommitCommitter
+ * @covers \Devboard\GitHub\Commit\GitHubCommitCommitter
  * @group  unit
  *
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
